backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
email-validator>=2.0.0
asyncpg>=0.28.0
psutil>=5.9.0
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6 